fix: load API keys correctly and persist properties from external sources
- Force-recreate containers (restart alone doesn't re-inject env_file
  vars); RENTCAST, FRED, and CENSUS keys are now live in the backend
- Fix _upsert_property: commit+refresh instead of flush so properties
  fetched from RentCast are actually persisted across requests
- Fix lat/lng schema type Decimal→float so Mapbox receives JSON numbers
  instead of strings (markers were silently dropping because setLngLat
  requires numeric coordinates)

+++ backend/app/schemas/property.py
@@ -85,8 +85,8 @@ class PropertyBase(BaseModel):
     state: str
     zip_code: str
     county: Optional[str] = None
-    lat: Optional[Decimal] = None
-    lng: Optional[Decimal] = None
+    lat: Optional[float] = None
+    lng: Optional[float] = None
     beds: Optional[int] = None
     baths: Optional[Decimal] = None
     sqft: Optional[int] = None
